// apps/mobile/src/api/gen/types/SessionDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type SessionDto = {
    /**
     * @type string, uuid
     */
    id: string;
    /**
     * @type string
     */
    deviceName: string;
    /**
     * @type integer, int64
     */
    createdAt: number;
    /**
     * @type integer, int64
     */
    lastUsedAt: number;
    /**
     * @type boolean
     */
    isCurrentDevice: boolean;
};